requirements_db.txt
click==8.0.4
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Werkzeug==2.0.3
numpy==1.22.0
scipy==1.8.0
matplotlib==3.3.1
psycopg2-binary==2.8.6
pandas==1.3.5